State-path seam descends to intendant-core
The merge of main brought PR #39's seam (intendant_home /
intendant_home_in) into platform.rs — which this branch's skills.rs,
now in intendant-core, cannot reach without a dependency cycle. The
seam family (home_dir, intendant_home, intendant_home_in, the override
parser) plus its six tests move to intendant-core::state_paths;
platform re-exports them at the old mount so every platform:: caller
keeps compiling, and skills.rs uses the crate-local path. The
cfg(test) scratch-home default travels with its tests, so the
protection surface is unchanged.
